BOOT: Add bootloader remembers boot-device
so one of the core requirements of bootloader is to store informations
that are struictly accessible just in real mode, there are multiple
information that could be stores but for this commit bootloader stores
boot_device information, and it is most crucial part of information that
is to be derived later regarding bootable device information, also
consditions for storing required type of strurcture has been created for
ease of use.

remember that `dl` stores drive device id and with refernce to
[table](https://en.wikipedia.org/wiki/INT_13H#:~:text=%5Bedit%5D-,Drive%20Table,-DL%20%3D%2000h)
it becomes easier to seperate them into floppy and hdrive category, this
information will later be used for acessing information regarding disk.

before switching to kernel main.